streamlit-app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
altair==5.5.0
    # via streamlit
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.2.1
    # via streamlit-authenticator
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
captcha==0.6.0
    # via streamlit-authenticator
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
colorama==0.4.6
    # via click
extra-streamlit-components==0.1.71
    # via streamlit-authenticator
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
idna==3.10
    # via requests
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.20.1
    # via altair
numpy==2.2.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.1.0
    # via
    #   captcha
    #   streamlit
protobuf==5.29.2
    # via streamlit
pyarrow==18.1.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyjwt==2.10.1
    # via streamlit-authenticator
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via streamlit-authenticator
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.41.1
    # via
    #   -r .\requirements.in
    #   extra-streamlit-components
    #   streamlit-authenticator
    #   streamlit-option-menu
streamlit-authenticator==0.4.1
    # via -r .\requirements.in
streamlit-option-menu==0.4.0
    # via -r .\requirements.in
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
